Move class Type and TypeCast to type.py
