+++ greenplumpython/__init__.py
@@ -1,8 +1,8 @@
 from greenplumpython.db import Database, database
-from greenplumpython.expr import get_type
 from greenplumpython.func import create_aggregate  # type: ignore
 from greenplumpython.func import create_array_function  # type: ignore
 from greenplumpython.func import create_function  # type: ignore
 from greenplumpython.func import aggregate, function
 from greenplumpython.op import operator
 from greenplumpython.table import Table, table, values
+from greenplumpython.type import get_type
